Extract AST processing to separate packages
Making it independent from the jsonnet package breaks the circular
dependency during stdast generation.
