+++ BUILD.bazel
@@ -13,13 +13,11 @@ go_library(
     name = "go_default_library",
     srcs = [
         "builtins.go",
-        "desugarer.go",
         "doc.go",
         "error_formatter.go",
         "imports.go",
         "interpreter.go",
         "runtime_error.go",
-        "static_analyzer.go",
         "thunks.go",
         "value.go",
         "vm.go",
@@ -28,6 +26,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//ast:go_default_library",
+        "//internal/transformations:go_default_library",
         "//parser:go_default_library",
     ],
 )
@@ -35,11 +34,9 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "desugarer_test.go",
         "interpreter_test.go",
         "jsonnet_test.go",
         "main_test.go",
-        "static_analyzer_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
