+++ c-bindings/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "internal.h",
         "json.cpp",
         "json.h",
+        "libgojsonnet.h",
         "libjsonnet.cpp",
     ],
     cdeps = [
