+++ cmd/dumpstdlibast/dumpstdlibast.go
@@ -19,8 +19,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/google/go-jsonnet"
-	"github.com/google/go-jsonnet/dump"
+	"github.com/google/go-jsonnet/internal/dump"
+	"github.com/google/go-jsonnet/internal/transformations"
 )
 
 func main() {
@@ -33,7 +33,7 @@ func main() {
 		panic(err)
 	}
 
-	node, err := jsonnet.SnippetToAST("<std>", string(buf))
+	node, err := transformations.SnippetToAST("<std>", string(buf))
 	if err != nil {
 		panic(err)
 	}
