+++ internal/dump/pointermap.go
