+++ internal/transformations/BUILD.bazel
@@ -1,0 +1,26 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "desugarer.go",
+        "static_analyzer.go",
+        "transformations.go",
+    ],
+    importpath = "github.com/google/go-jsonnet/internal/transformations",
+    visibility = ["//:__subpackages__"],
+    deps = [
+        "//ast:go_default_library",
+        "//parser:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "desugarer_test.go",
+        "static_analyzer_test.go",
+    ],
+    embed = [":go_default_library"],
+    deps = ["//ast:go_default_library"],
+)
